refactor: rename project to music_player
Renames the project from my_discord_bot to music_player.
This includes renaming the package, entrypoint, and updating
the README.
