+++ src/music_player/cogs/__init__.py
@@ -1,0 +1,4 @@
+from music_player.cogs.bot_cog import BotCog
+from music_player.cogs.music import Music
+
+COGS: list[type[BotCog]] = [Music]
