+++ src/music_player/exceptions.py
