--- src/my_discord_bot/__init__.py
@@ -1,6 +1,0 @@
-from my_discord_bot.bot import MyBot
-from my_discord_bot.settings import DISCORD_TOKEN
-
-
-def main() -> None:
-    MyBot().run(DISCORD_TOKEN)
