--- src/my_discord_bot/cogs/__init__.py
@@ -1,4 +1,0 @@
-from my_discord_bot.cogs.bot_cog import BotCog
-from my_discord_bot.cogs.music import Music
-
-COGS: list[type[BotCog]] = [Music]
